Compilation fixes

+++ haskell/maths/src/primefun/semiprime.hs
@@ -4,7 +4,7 @@ import Data.Numbers.Primes
 
 main :: IO ()
 main = do
-    print [s `mod` (2 ^ n) | n <- [1..10]]
+    print [s `mod` (2 ^ n) | n <- [1..10::Integer]]
     
 
 p1, p2, s :: Integer
